Add files via upload
A more complicated version with line representing synchronization between the bugs and rings that show the amount of activity in the neurons. If there is a lot of neurons, it does not look good. But with 16 it is nice to look at the different activity levels of the bug brains. Also, there is brain coupler directly trying to make the bug brains to tune the bug brain to have similar eeg activity as the model.